+++ net/http_net.hpp
@@ -135,7 +135,8 @@ thread_base("http"), use_ssl(server.size() >= 5 && server.substr(0, 5) == "https
 	server_port = (use_ssl ? 443 : 80);
 	const size_t colon_pos = server_name.find(":");
 	if(colon_pos != string::npos) {
-		server_port = strtoul(server_name.substr(colon_pos + 1, server_name.size() - colon_pos - 1).c_str(), nullptr, 10);
+		server_port = (unsigned short int)strtoul(server_name.substr(colon_pos + 1, server_name.size() - colon_pos - 1).c_str(),
+												  nullptr, 10);
 		server_name = server_name.substr(0, colon_pos);
 	}
